9971 Make getrandom(2) a public interface Reviewed by: Dan McDonald <danmcd@joyent.com> Reviewed by: Mike Gerdts <mike.gerdts@joyent.com> Reviewed by: Peter Tribble <peter.tribble@gmail.com> Reviewed by: Robert Mustacchi <rm@joyent.com> Reviewed by: Andy Fiddaman <omnios@citrus-it.net> Reviewed by: Igor Kozhukhov <igor@dilos.org> Approved by: Joshua M. Clulow <josh@sysmgr.org>

+++ usr/src/lib/libc/port/sys/getrandom.c
@@ -10,15 +10,15 @@
  */
 
 /*
- * Copyright (c) 2015 Joyent, Inc.
+ * Copyright (c) 2018 Joyent, Inc.
  */
 
 #include <sys/types.h>
 #include <sys/syscall.h>
 #include <sys/random.h>
 
-int
-getrandom(void *buf, size_t len, int flags)
+ssize_t
+getrandom(void *buf, size_t len, unsigned int flags)
 {
 	return (syscall(SYS_getrandom, buf, len, flags));
 }
